app-client/requirements.txt
python-firebase
roslibpy
python-dotenv
sshtunnel
pyinstaller